requirements.txt
streamlit
crewai
langchain
chromadb
pymupdf
pandas
python-dotenv
sentence-transformers
langchain-text-splitters